pyproject.toml
[tool.poetry]
name = "avatara_gendiff"
version = "1.2.1"
description = "Json and Yaml files difference generator"
authors = ["Andrey Larin <larin.projects@gmail.com>"]
packages = [
  { include = "gendiff" },
]

[tool.poetry.dependencies]
python = "^3.7"
PyYAML = "^5.3.1"

[tool.poetry.dev-dependencies]
flake8 = "^3.8.3"
pytest = "^5.4.3"
pytest-cov = "^2.10.1"

[tool.poetry.scripts]
gendiff = "gendiff.scripts.gendiff:main"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
